requirements.txt
discord.py==2.5.2
tabulate==0.8.10
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
mysql==0.0.3
mysql-connector-python==9.4.0
mysqlclient==2.1.1
propcache==0.3.2
python-dotenv==1.1.1
typing_extensions==4.14.1
yarl==1.20.1
